# docs/API_OPENAPI.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BehavioralWell \u2014 Web + Android Shared Platform",
    "description": "Privacy-conscious multimodal behavioral risk detection, digital phenotyping, and early intervention API engine.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/consent/{user_id}": {
      "get": {
        "tags": [
          "Consent & Privacy"
        ],
        "summary": "Get Consent",
        "operationId": "get_consent_api_v1_consent__user_id__get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Consent & Privacy"
        ],
        "summary": "Update Consent",
        "operationId": "update_consent_api_v1_consent__user_id__put",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsentUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/interventions/{user_id}/recommendations": {
      "get": {
        "tags": [
          "Interventions & Check-ins"
        ],
        "summary": "Get Recommendations",
        "operationId": "get_recommendations_api_v1_interventions__user_id__recommendations_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/interventions/{user_id}/self-check": {
      "post": {
        "tags": [
          "Interventions & Check-ins"
        ],
        "summary": "Submit Self Check",
        "operationId": "submit_self_check_api_v1_interventions__user_id__self_check_post",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelfReportInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/interventions/{user_id}/start": {
      "post": {
        "tags": [
          "Interventions & Check-ins"
        ],
        "summary": "Start Intervention",
        "operationId": "start_intervention_api_v1_interventions__user_id__start_post",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InterventionStartInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/interventions/sessions/{session_id}/complete": {
      "post": {
        "tags": [
          "Interventions & Check-ins"
        ],
        "summary": "Complete Intervention",
        "operationId": "complete_intervention_api_v1_interventions_sessions__session_id__complete_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InterventionCompleteInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/consent/{user_id}": {
      "get": {
        "tags": [
          "Consent & Privacy"
        ],
        "summary": "Get Consent",
        "operationId": "get_consent_api_consent__user_id__get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Consent & Privacy"
        ],
        "summary": "Update Consent",
        "operationId": "update_consent_api_consent__user_id__put",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsentUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/interventions/{user_id}/recommendations": {
      "get": {
        "tags": [
          "Interventions & Check-ins"
        ],
        "summary": "Get Recommendations",
        "operationId": "get_recommendations_api_interventions__user_id__recommendations_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/interventions/{user_id}/self-check": {
      "post": {
        "tags": [
          "Interventions & Check-ins"
        ],
        "summary": "Submit Self Check",
        "operationId": "submit_self_check_api_interventions__user_id__self_check_post",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelfReportInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/interventions/{user_id}/start": {
      "post": {
        "tags": [
          "Interventions & Check-ins"
        ],
        "summary": "Start Intervention",
        "operationId": "start_intervention_api_interventions__user_id__start_post",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InterventionStartInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/interventions/sessions/{session_id}/complete": {
      "post": {
        "tags": [
          "Interventions & Check-ins"
        ],
        "summary": "Complete Intervention",
        "operationId": "complete_intervention_api_interventions_sessions__session_id__complete_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InterventionCompleteInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConsentResponse": {
        "properties": {
          "keyboard_enabled": {
            "type": "boolean",
            "title": "Keyboard Enabled"
          },
          "usage_enabled": {
            "type": "boolean",
            "title": "Usage Enabled"
          },
          "motion_enabled": {
            "type": "boolean",
            "title": "Motion Enabled"
          },
          "work_enabled": {
            "type": "boolean",
            "title": "Work Enabled"
          },
          "mobility_enabled": {
            "type": "boolean",
            "title": "Mobility Enabled"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "keyboard_enabled",
          "usage_enabled",
          "motion_enabled",
          "work_enabled",
          "mobility_enabled",
          "user_id",
          "updated_at"
        ],
        "title": "ConsentResponse"
      },
      "ConsentUpdate": {
        "properties": {
          "keyboard_enabled": {
            "type": "boolean",
            "title": "Keyboard Enabled"
          },
          "usage_enabled": {
            "type": "boolean",
            "title": "Usage Enabled"
          },
          "motion_enabled": {
            "type": "boolean",
            "title": "Motion Enabled"
          },
          "work_enabled": {
            "type": "boolean",
            "title": "Work Enabled"
          },
          "mobility_enabled": {
            "type": "boolean",
            "title": "Mobility Enabled"
          }
        },
        "type": "object",
        "required": [
          "keyboard_enabled",
          "usage_enabled",
          "motion_enabled",
          "work_enabled",
          "mobility_enabled"
        ],
        "title": "ConsentUpdate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InterventionCompleteInput": {
        "properties": {
          "feedback_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feedback Score"
          },
          "result_metrics": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Metrics"
          }
        },
        "type": "object",
        "title": "InterventionCompleteInput"
      },
      "InterventionStartInput": {
        "properties": {
          "activity_type": {
            "type": "string",
            "title": "Activity Type"
          }
        },
        "type": "object",
        "required": [
          "activity_type"
        ],
        "title": "InterventionStartInput"
      },
...
      "SelfReportInput": {
        "properties": {
          "mood": {
            "type": "string",
            "title": "Mood"
          },
          "stress_level": {
            "type": "integer",
            "title": "Stress Level"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "mood",
          "stress_level"
        ],
        "title": "SelfReportInput"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}